+++ src/operator/numpy/np_where_forward_op.cu
@@ -19,30 +19,22 @@
 
 /*!
  * Copyright (c) 2017 by Contributors
- * \file np_where_op.cu
+ * \file np_where_forward_op.cu
  * \brief GPU Implementation of numpy operator where
  */
 
-#include "np_where_op-inl.h"
+#include "np_where_forward_op-inl.h"
 
 namespace mxnet {
 namespace op {
 
 NNVM_REGISTER_OP(_npi_where).set_attr<FCompute>("FCompute<gpu>", NumpyWhereOpForward<gpu>);
 
-NNVM_REGISTER_OP(_backward_np_where).set_attr<FCompute>("FCompute<gpu>", NumpyWhereOpBackward<gpu>);
-
 NNVM_REGISTER_OP(_npi_where_lscalar)
     .set_attr<FCompute>("FCompute<gpu>", NumpyWhereScalarOpForward<gpu, true>);
 
 NNVM_REGISTER_OP(_npi_where_rscalar)
     .set_attr<FCompute>("FCompute<gpu>", NumpyWhereScalarOpForward<gpu, false>);
-
-NNVM_REGISTER_OP(_backward_np_where_lscalar)
-    .set_attr<FCompute>("FCompute<gpu>", NumpyWhereScalarOpBackward<gpu, true>);
-
-NNVM_REGISTER_OP(_backward_np_where_rscalar)
-    .set_attr<FCompute>("FCompute<gpu>", NumpyWhereScalarOpBackward<gpu, false>);
 
 NNVM_REGISTER_OP(_npi_where_scalar2)
     .set_attr<FCompute>("FCompute<gpu>", NumpyWhereScalar2OpForward<gpu>);
